+++ src/extensions/exception-autocapture/error-conversion.ts
@@ -10,9 +10,45 @@ import {
 } from './type-checking'
 import { defaultStackParser, StackFrame } from './stack-trace'
 
-import { isEmptyString, isNumber, isString, isUndefined } from '../../utils/type-utils'
-import { ErrorEventArgs, ErrorProperties, SeverityLevel, severityLevels } from '../../types'
+import { isEmptyString, isString, isUndefined } from '../../utils/type-utils'
+import { ErrorEventArgs, ErrorMetadata, SeverityLevel, severityLevels } from '../../types'
 
+export interface ErrorProperties {
+    $exception_list: Exception[]
+    $exception_level?: SeverityLevel
+    $exception_DOMException_code?: string
+    $exception_personURL?: string
+}
+
+export interface Exception {
+    type?: string
+    value?: string
+    mechanism?: {
+        /**
+         * In theory, whether or not the exception has been handled by the user. In practice, whether or not we see it before
+         * it hits the global error/rejection handlers, whether through explicit handling by the user or auto instrumentation.
+         */
+        handled?: boolean
+        type?: string
+        source?: string
+        /**
+         * True when `captureException` is called with anything other than an instance of `Error` (or, in the case of browser,
+         * an instance of `ErrorEvent`, `DOMError`, or `DOMException`). causing us to create a synthetic error in an attempt
+         * to recreate the stacktrace.
+         */
+        synthetic?: boolean
+    }
+    module?: string
+    thread_id?: number
+    stacktrace?: {
+        frames?: StackFrame[]
+    }
+}
+
+export interface ErrorConversions {
+    errorToProperties: (args: ErrorEventArgs, metadata?: ErrorMetadata) => ErrorProperties
+    unhandledRejectionToProperties: (args: [ev: PromiseRejectionEvent]) => ErrorProperties
+}
 /**
  * based on the very wonderful MIT licensed Sentry SDK
  */
@@ -20,32 +56,18 @@ import { ErrorEventArgs, ErrorProperties, SeverityLevel, severityLevels } from '
 const ERROR_TYPES_PATTERN =
     /^(?:[Uu]ncaught (?:exception: )?)?(?:((?:Eval|Internal|Range|Reference|Syntax|Type|URI|)Error): )?(.*)$/i
 
-const reactMinifiedRegexp = /Minified React error #\d+;/i
-
-function getPopSize(ex: Error & { framesToPop?: number }): number {
-    if (ex) {
-        if (isNumber(ex.framesToPop)) {
-            return ex.framesToPop
-        }
-
-        if (reactMinifiedRegexp.test(ex.message)) {
-            return 1
-        }
-    }
-
-    return 0
-}
-
-export function parseStackFrames(ex: Error & { framesToPop?: number; stacktrace?: string }): StackFrame[] {
+export function parseStackFrames(ex: Error & { stacktrace?: string }, framesToPop: number = 0): StackFrame[] {
     // Access and store the stacktrace property before doing ANYTHING
     // else to it because Opera is not very good at providing it
     // reliably in other circumstances.
     const stacktrace = ex.stacktrace || ex.stack || ''
 
-    const popSize = getPopSize(ex)
+    const skipLines = getSkipFirstStackStringLines(ex)
 
     try {
-        return defaultStackParser(stacktrace, popSize)
+        const frames = defaultStackParser(stacktrace, skipLines)
+        // frames are reversed so we remove the from the back of the array
+        return frames.slice(0, frames.length - framesToPop)
     } catch {
         // no-empty
     }
@@ -53,21 +75,99 @@ export function parseStackFrames(ex: Error & { framesToPop?: number; stacktrace?
     return []
 }
 
-function errorPropertiesFromError(error: Error): ErrorProperties {
+const reactMinifiedRegexp = /Minified React error #\d+;/i
+
+/**
+ * Certain known React errors contain links that would be falsely
+ * parsed as frames. This function check for these errors and
+ * returns number of the stack string lines to skip.
+ */
+function getSkipFirstStackStringLines(ex: Error): number {
+    if (ex && reactMinifiedRegexp.test(ex.message)) {
+        return 1
+    }
+
+    return 0
+}
+
+function errorPropertiesFromError(error: Error, metadata?: ErrorMetadata): ErrorProperties {
     const frames = parseStackFrames(error)
 
+    const handled = metadata?.handled ?? true
+    const synthetic = metadata?.synthetic ?? false
+
+    const exceptionType = metadata?.overrideExceptionType ? metadata.overrideExceptionType : error.name
+    const exceptionMessage = metadata?.overrideExceptionMessage
+        ? metadata.overrideExceptionMessage
+        : extractMessage(error)
+
     return {
-        $exception_type: error.name,
-        $exception_message: error.message,
-        $exception_stack_trace_raw: JSON.stringify(frames),
+        $exception_list: [
+            {
+                type: exceptionType,
+                value: exceptionMessage,
+                stacktrace: {
+                    frames,
+                },
+                mechanism: {
+                    handled,
+                    synthetic,
+                },
+            },
+        ],
         $exception_level: 'error',
     }
 }
 
-function errorPropertiesFromString(candidate: string): ErrorProperties {
+/**
+ * There are cases where stacktrace.message is an Event object
+ * https://github.com/getsentry/sentry-javascript/issues/1949
+ * In this specific case we try to extract stacktrace.message.error.message
+ */
+export function extractMessage(err: Error & { message: { error?: Error } }): string {
+    const message = err.message
+
+    if (message.error && typeof message.error.message === 'string') {
+        return message.error.message
+    }
+
+    return message
+}
+
+function errorPropertiesFromString(candidate: string, metadata?: ErrorMetadata): ErrorProperties {
+    // Defaults for metadata are based on what the error candidate is.
+    const handled = metadata?.handled ?? true
+    const synthetic = metadata?.synthetic ?? true
+
+    const exceptionType = metadata?.overrideExceptionType
+        ? metadata.overrideExceptionType
+        : metadata?.defaultExceptionType ?? 'Error'
+    const exceptionMessage = metadata?.overrideExceptionMessage
+        ? metadata.overrideExceptionMessage
+        : candidate
+        ? candidate
+        : metadata?.defaultExceptionMessage
+
+    const exception: Exception = {
+        type: exceptionType,
+        value: exceptionMessage,
+        mechanism: {
+            handled,
+            synthetic,
+        },
+    }
+
+    if (metadata?.syntheticException) {
+        // Kludge: strip the last frame from a synthetically created error
+        // so that it does not appear in a users stack trace
+        const frames = parseStackFrames(metadata.syntheticException, 1)
+        if (frames.length) {
+            exception.stacktrace = { frames }
+        }
+    }
+
     return {
-        $exception_type: 'Error',
-        $exception_message: candidate,
+        $exception_list: [exception],
         $exception_level: 'error',
     }
 }
@@ -103,35 +203,50 @@ function isSeverityLevel(x: unknown): x is SeverityLevel {
     return isString(x) && !isEmptyString(x) && severityLevels.indexOf(x as SeverityLevel) >= 0
 }
 
-function errorPropertiesFromObject(candidate: Record<string, unknown>): ErrorProperties {
+function errorPropertiesFromObject(candidate: Record<string, unknown>, metadata?: ErrorMetadata): ErrorProperties {
+    // Defaults for metadata are based on what the error candidate is.
+    const handled = metadata?.handled ?? true
+    const synthetic = metadata?.synthetic ?? true
+
+    const exceptionType = metadata?.overrideExceptionType
+        ? metadata.overrideExceptionType
+        : isEvent(candidate)
+        ? candidate.constructor.name
+        : 'Error'
+    const exceptionMessage = metadata?.overrideExceptionMessage
+        ? metadata.overrideExceptionMessage
+        : `Non-Error ${'exception'} captured with keys: ${extractExceptionKeysForMessage(candidate)}`
+
+    const exception: Exception = {
+        type: exceptionType,
+        value: exceptionMessage,
+        mechanism: {
+            handled,
+            synthetic,
+        },
+    }
+
+    if (metadata?.syntheticException) {
+        // Kludge: strip the last frame from a synthetically created error
+        // so that it does not appear in a users stack trace
+        const frames = parseStackFrames(metadata?.syntheticException, 1)
+        if (frames.length) {
+            exception.stacktrace = { frames }
+        }
+    }
+
     return {
-        $exception_type: isEvent(candidate) ? candidate.constructor.name : 'Error',
-        $exception_message: `Non-Error ${'exception'} captured with keys: ${extractExceptionKeysForMessage(candidate)}`,
+        $exception_list: [exception],
         $exception_level: isSeverityLevel(candidate.level) ? candidate.level : 'error',
     }
 }
 
-export function errorToProperties([event, source, lineno, colno, error]: ErrorEventArgs): ErrorProperties {
-    // some properties are not optional but, it's useful to start off without them enforced
-    let errorProperties: Omit<ErrorProperties, '$exception_type' | '$exception_message' | '$exception_level'> & {
-        $exception_type?: string
-        $exception_message?: string
-        $exception_level?: string
-    } = {}
-
-    if (isUndefined(error) && isString(event)) {
-        let name = 'Error'
-        let message = event
-        const groups = event.match(ERROR_TYPES_PATTERN)
-        if (groups) {
-            name = groups[1]
-            message = groups[2]
-        }
-        errorProperties = {
-            $exception_type: name,
-            $exception_message: message,
-        }
-    }
+export function errorToProperties(
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    [event, _, __, ___, error]: ErrorEventArgs,
+    metadata?: ErrorMetadata
+): ErrorProperties {
+    let errorProperties: ErrorProperties = { $exception_list: [] }
 
     const candidate = error || event
 
@@ -142,95 +257,92 @@ export function errorToProperties([event, source, lineno, colno, error]: ErrorEv
         const domException = candidate as unknown as DOMException
 
         if (isErrorWithStack(candidate)) {
-            errorProperties = errorPropertiesFromError(candidate as Error)
+            errorProperties = errorPropertiesFromError(candidate as Error, metadata)
         } else {
             const name = domException.name || (isDOMError(domException) ? 'DOMError' : 'DOMException')
             const message = domException.message ? `${name}: ${domException.message}` : name
-            errorProperties = errorPropertiesFromString(message)
-            errorProperties.$exception_type = isDOMError(domException) ? 'DOMError' : 'DOMException'
-            errorProperties.$exception_message = errorProperties.$exception_message || message
+            const exceptionType = isDOMError(domException) ? 'DOMError' : 'DOMException'
+            errorProperties = errorPropertiesFromString(message, {
+                ...metadata,
+                overrideExceptionType: exceptionType,
+                defaultExceptionMessage: message,
+            })
         }
         if ('code' in domException) {
             errorProperties['$exception_DOMException_code'] = `${domException.code}`
         }
+        return errorProperties
     } else if (isErrorEvent(candidate as ErrorEvent) && (candidate as ErrorEvent).error) {
-        errorProperties = errorPropertiesFromError((candidate as ErrorEvent).error as Error)
+        return errorPropertiesFromError((candidate as ErrorEvent).error as Error, metadata)
     } else if (isError(candidate)) {
-        errorProperties = errorPropertiesFromError(candidate)
+        return errorPropertiesFromError(candidate, metadata)
     } else if (isPlainObject(candidate) || isEvent(candidate)) {
         // group these by using the keys available on the object
         const objectException = candidate as Record<string, unknown>
-        errorProperties = errorPropertiesFromObject(objectException)
-        errorProperties.$exception_is_synthetic = true
-    } else {
-        // If none of previous checks were valid, then it must be a string
-        errorProperties.$exception_type = errorProperties.$exception_type || 'Error'
-        errorProperties.$exception_message = errorProperties.$exception_message || candidate
-        errorProperties.$exception_is_synthetic = true
-    }
+        return errorPropertiesFromObject(objectException, metadata)
+    } else if (isUndefined(error) && isString(event)) {
+        let name = 'Error'
+        let message = event
+        const groups = event.match(ERROR_TYPES_PATTERN)
+        if (groups) {
+            name = groups[1]
+            message = groups[2]
+        }
 
-    return {
-        ...errorProperties,
-        // now we make sure the mandatory fields that were made optional are present
-        $exception_type: errorProperties.$exception_type || 'UnknownErrorType',
-        $exception_message: errorProperties.$exception_message || '',
-        $exception_level: isSeverityLevel(errorProperties.$exception_level)
-            ? errorProperties.$exception_level
-            : 'error',
-        ...(source
-            ? {
-                  $exception_source: source, // TODO get this from URL if not present
-              }
-            : {}),
-        ...(lineno ? { $exception_lineno: lineno } : {}),
-        ...(colno ? { $exception_colno: colno } : {}),
+        return errorPropertiesFromString(message, {
+            ...metadata,
+            overrideExceptionType: name,
+            defaultExceptionMessage: message,
+        })
+    } else {
+        return errorPropertiesFromString(candidate as string, metadata)
     }
 }
 
 export function unhandledRejectionToProperties([ev]: [ev: PromiseRejectionEvent]): ErrorProperties {
+    const error = getUnhandledRejectionError(ev)
+
+    if (isPrimitive(error)) {
+        return errorPropertiesFromString(`Non-Error promise rejection captured with value: ${String(error)}`, {
+            handled: false,
+            synthetic: false,
+            overrideExceptionType: 'UnhandledRejection',
+        })
+    }
+
+    return errorToProperties([error as string | Event], {
+        handled: false,
+        overrideExceptionType: 'UnhandledRejection',
+        defaultExceptionMessage: String(error),
+    })
+}
+
+function getUnhandledRejectionError(error: unknown): unknown {
+    if (isPrimitive(error)) {
+        return error
+    }
+
     // dig the object of the rejection out of known event types
-    let error: unknown = ev
     try {
+        type ErrorWithReason = { reason: unknown }
         // PromiseRejectionEvents store the object of the rejection under 'reason'
         // see https://developer.mozilla.org/en-US/docs/Web/API/PromiseRejectionEvent
-        if ('reason' in ev) {
-            error = ev.reason
+        if ('reason' in (error as ErrorWithReason)) {
+            return (error as ErrorWithReason).reason
         }
+
+        type CustomEventWithDetail = { detail: { reason: unknown } }
         // something, somewhere, (likely a browser extension) effectively casts PromiseRejectionEvents
         // to CustomEvents, moving the `promise` and `reason` attributes of the PRE into
         // the CustomEvent's `detail` attribute, since they're not part of CustomEvent's spec
         // see https://developer.mozilla.org/en-US/docs/Web/API/CustomEvent and
         // https://github.com/getsentry/sentry-javascript/issues/2380
-        else if ('detail' in ev && 'reason' in (ev as any).detail) {
-            error = (ev as any).detail.reason
+        if ('detail' in (error as CustomEventWithDetail) && 'reason' in (error as CustomEventWithDetail).detail) {
+            return (error as CustomEventWithDetail).detail.reason
         }
     } catch {
         // no-empty
     }
 
-    // some properties are not optional but, it's useful to start off without them enforced
-    let errorProperties: Omit<ErrorProperties, '$exception_type' | '$exception_message' | '$exception_level'> & {
-        $exception_type?: string
-        $exception_message?: string
-        $exception_level?: string
-    } = {}
-    if (isPrimitive(error)) {
-        errorProperties = {
-            $exception_message: `Non-Error promise rejection captured with value: ${String(error)}`,
-        }
-    } else {
-        errorProperties = errorToProperties([error as string | Event])
-    }
-    errorProperties.$exception_handled = false
-
-    return {
-        ...errorProperties,
-        // now we make sure the mandatory fields that were made optional are present
-        $exception_type: (errorProperties.$exception_type = 'UnhandledRejection'),
-        $exception_message: (errorProperties.$exception_message =
-            errorProperties.$exception_message || (ev as any).reason || String(error)),
-        $exception_level: isSeverityLevel(errorProperties.$exception_level)
-            ? errorProperties.$exception_level
-            : 'error',
-    }
+    return error
 }