+++ src/extensions/replay/sessionrecording.ts
@@ -43,6 +43,14 @@ import { isLocalhost } from '../../utils/request-utils'
 import { MutationRateLimiter } from './mutation-rate-limiter'
 import { gzipSync, strFromU8, strToU8 } from 'fflate'
 
+type SessionStartReason =
+    | 'sampling_override'
+    | 'recording_initialized'
+    | 'linked_flag_match'
+    | 'linked_flag_override'
+    | 'sampling'
+    | 'session_id_changed'
+
 const BASE_ENDPOINT = '/s/'
 
 const ONE_MINUTE = 1000 * 60
@@ -443,9 +451,9 @@ export class SessionRecording {
         }
     }
 
-    startIfEnabledOrStop() {
+    startIfEnabledOrStop(startReason?: SessionStartReason) {
         if (this.isRecordingEnabled) {
-            this._startCapture()
+            this._startCapture(startReason)
 
             // calling addEventListener multiple times is safe and will not add duplicates
             window?.addEventListener('beforeunload', this._onBeforeUnload)
@@ -550,15 +558,21 @@ export class SessionRecording {
             shouldSample = storedIsSampled
         }
 
-        if (!shouldSample && makeDecision) {
-            logger.warn(
-                LOGGER_PREFIX +
-                    ` Sample rate (${currentSampleRate}) has determined that this sessionId (${sessionId}) will not be sent to the server.`
-            )
+        if (makeDecision) {
+            if (shouldSample) {
+                this._reportStarted('sampling')
+            } else {
+                logger.warn(
+                    LOGGER_PREFIX +
+                        ` Sample rate (${currentSampleRate}) has determined that this sessionId (${sessionId}) will not be sent to the server.`
+                )
+            }
+
+            this._tryAddCustomEvent('samplingDecisionMade', {
+                sampleRate: currentSampleRate,
+                isSampled: shouldSample,
+            })
         }
-        this._tryAddCustomEvent('samplingDecisionMade', {
-            sampleRate: currentSampleRate,
-        })
 
         this.instance.persistence?.register({
             [SESSION_RECORDING_IS_SAMPLED]: shouldSample,
@@ -590,6 +604,7 @@ export class SessionRecording {
                     const tag = 'linked flag matched'
                     logger.info(LOGGER_PREFIX + ' ' + tag, payload)
                     this._tryAddCustomEvent(tag, payload)
+                    this._reportStarted('linked_flag_match')
                 }
                 this._linkedFlagSeen = linkedFlagMatches
             })
@@ -667,14 +682,14 @@ export class SessionRecording {
         })
     }
 
-    private _startCapture() {
-        if (isUndefined(Object.assign)) {
+    private _startCapture(startReason?: SessionStartReason) {
+        if (isUndefined(Object.assign) || isUndefined(Array.from)) {
             // According to the rrweb docs, rrweb is not supported on IE11 and below:
             // "rrweb does not support IE11 and below because it uses the MutationObserver API which was supported by these browsers."
             // https://github.com/rrweb-io/rrweb/blob/master/guide.md#compatibility-note
             //
             // However, MutationObserver does exist on IE11, it just doesn't work well and does not detect all changes.
-            // Instead, when we load "recorder.js", the first JS error is about "Object.assign" being undefined.
+            // Instead, when we load "recorder.js", the first JS error is about "Object.assign" and "Array.from" being undefined.
             // Thus instead of MutationObserver, we look for this function and block recording if it's undefined.
             return
         }
@@ -705,6 +720,11 @@ export class SessionRecording {
             })
         } else {
             this._onScriptLoaded()
+        }
+
+        logger.info(LOGGER_PREFIX + ' starting')
+        if (this.status === 'active') {
+            this._reportStarted(startReason || 'recording_initialized')
         }
     }
 
@@ -779,7 +799,7 @@ export class SessionRecording {
 
         if (sessionIdChanged || windowIdChanged) {
             this.stopRecording()
-            this.startIfEnabledOrStop()
+            this.startIfEnabledOrStop('session_id_changed')
         } else if (returningFromIdle) {
             this._scheduleFullSnapshot()
         }
@@ -895,11 +915,6 @@ export class SessionRecording {
 
         this._tryAddCustomEvent('$posthog_config', {
             config: this.instance.config,
-        })
-
-        logger.info(LOGGER_PREFIX + ' started', {
-            idleThreshold: this.sessionIdleThresholdMilliseconds,
-            maxIdleTime: this.sessionManager.sessionTimeoutMs,
         })
     }
 
@@ -1192,5 +1207,29 @@ export class SessionRecording {
      * */
     public overrideLinkedFlag() {
         this._linkedFlagSeen = true
+        this._reportStarted('linked_flag_override')
+    }
+
+    /**
+     * this ignores the sampling config and causes capture to start
+     * (if recording would have started had the flag been received i.e. it does not override other config).
+     *
+     * It is not usual to call this directly,
+     * instead call `posthog.startSessionRecording({sampling: true})`
+     * */
+    public overrideSampling() {
+        this.instance.persistence?.register({
+            // short-circuits the `makeSamplingDecision` function in the session recording module
+            [SESSION_RECORDING_IS_SAMPLED]: true,
+        })
+        this._reportStarted('sampling_override')
+    }
+
+    private _reportStarted(startReason: SessionStartReason, shouldReport: () => boolean = () => true) {
+        if (shouldReport()) {
+            this.instance.register_for_session({
+                $session_recording_start_reason: startReason,
+            })
+        }
     }
 }